Added the basic counters needed for each shard, plus some mechanism to differentiate between a client and server to write to distinct MMV files for local testing (e.g. 'elasticsearch-server.mmv' to stop clashes.
Still having some issue with clashing though, when a subsequent client app starts up and writes to -client.mmv, the pcp metrics are getting 'lost' on the server mmv for some reason.

Also discovered a case when an index is deleted, but then re-added, then the metric will be attempted to be re-added, which is a no no, need Cowan's merge to ensure re-use of a Metric if it's needed again.